+++ examples/t.py
@@ -1,0 +1,13 @@
+
+def testsum():
+    t = sum([1, 2, 3])
+    print(t)
+
+def test_sum_tuple():
+    t = sum((1, 2, 2))
+    print(t)
+
+testsum()
+test_sum_tuple()
+
+print("Everything passed")